Restructured. Refined significantly.
