+++ typings/.gitignore
@@ -1,0 +1,8 @@
+# generated documentation
+doc
+
+# dependencies used to test azure-mobile-apps.d.ts
+express
+mime
+node
+serve-static
